% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{ae}
\alias{ae}
\title{Absolute Error}
\usage{
ae(y, x, na.action = na.omit, aggregate = FALSE, ...)
}
\arguments{
\item{y}{vector of observations}

\item{x}{matrix of samples of a predictive distribution or vector containing the medians of a predictive distribution (depending on \code{y}; see details)}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}

\item{aggregate}{logical or function for aggregating the single scores, e.g. \code{sum}, \code{mean}, \code{weighted.mean}, ....
Default: \code{FALSE}, i.e. no aggregation function.}

\item{...}{further arguments passed to the \code{aggregate} function.}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Absolute Error (AE) given observations of a univariate variable and samples of a predictive distribution.
}
\details{
For a vector \code{y} of length n, \code{x} can be given as matrix of samples of a predictive distribution
with n rows, where the i-th entry of \code{y} belongs to the i-th row
of \code{x}. The columns of \code{x} represent the samples of a predictive distribution.
The row-wise medians are determined by its sample version.

If the medians of a predictive distribution are directly
available, \code{x} can be given as vector containing the medians, where
the i-th entry of \code{y} belongs to the i-th entry of \code{x}.

In addition, with this function, the Mean Absolute Error (MAE) can be calculated by specifying \code{aggregate = mean}.

A lower AE indicates a better forecast.
}
\examples{
# simulated data
n <- 30
m <- 50
y <- rnorm(n)
x1 <- matrix(rnorm(n*m), ncol = m)
x2 <- apply(x1, 1, median)

# ae calculation
ae(y = y, x = x1)
ae(y = y, x = x1, aggregate = mean)

ae(y = y, x = x2)
(mae <- ae(y = y, x = x2, aggregate = mean))

}
\references{
Gneiting, T. (2011). Making and Evaluating Point Forecasts. Journal of the American Statistical Association, 106(494), 746-762.
}
\author{
David Jobst
}
