% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh.test.R
\name{bh.test}
\alias{bh.test}
\title{Benjamini-Hochberg-Procedure}
\usage{
bh.test(p, alpha = 0.05, na.action = na.omit)
}
\arguments{
\item{p}{vector of p-values on which the Benjamini-Hochberg-Procedure should be applied}

\item{alpha}{numeric; false discovery rate at level alpha; alpha is number between 0 and 1}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}
}
\value{
A logical vector where the value \code{TRUE} indicates that the p-value at the
corresponding  position is significant after the Benjamini-Hochberg-Procedure
with false discovery rate \code{alpha}. If the value is \code{FALSE} the p-value at the
corresponding  position is not significant after the
Benjamini-Hochberg-Procedure with false discovery rate \code{alpha}.
}
\description{
This function performs the Benjamini-Hochberg-Procedure for different p-values in the multiple testing setting.
This function computes the significance of the individual tests.
}
\examples{
# simulated data
p <- runif(100, 0, 0.06)

# Benjamini-Hochberg-Procedure
bh.test(p = p)
bh.test(p = p, alpha = 0.1)


}
\references{
Benjamini, Y. and Hochberg, Y. (1995). Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing. Journal of the Royal Statistical Society, 57, 289-300.
}
\author{
David Jobst
}
