% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.plot.R
\name{box.plot}
\alias{box.plot}
\title{Box Plot}
\usage{
box.plot(
  x,
  names,
  colors = NULL,
  x.lab = "",
  y.lab = "",
  title = "Box Plot",
  legend = "right"
)
}
\arguments{
\item{x}{matrix of given dataset}

\item{names}{character; column names of \code{x} (depending on \code{x}; see details)}

\item{colors}{character; color for each object in \code{names}; default: ggplot2 setting (depending on \code{x}; see details)}

\item{x.lab}{character; label of x-axis; default: ""}

\item{y.lab}{character; label of y-axis; default: ""}

\item{title}{character; title of the plot; default: "\code{Box Plot}"}

\item{legend}{character; position of the legend; default: "\code{right}" (see details)}
}
\value{
ggplot object with box plots for a given dataset.
}
\description{
This function creates box plots for a given dataset.
}
\details{
For matrix \code{x}, each column belongs to one object in \code{names} and \code{colors}.
If only one value is provided for \code{names} and \code{colors} this value is
used for all column entries in \code{x}.
Only finite values of \code{x} are used.

For \code{colors} the specifications of \code{color} in ggplot2 are used.

For the parameter \code{legend} the following options are available: "\code{right}", "\code{left}", "\code{bottom}", "\code{top}", "\code{hide}".
}
\examples{
#simulated data
n <- 30
x <- cbind(rnorm(n), rgamma(n, shape = 1))
names <- c("A", "B")
colors <- c("darkred", "steelblue")
x.lab <- "Models"
y.lab <- "Values"
title <- "Box Plots"
legend <- "top"

#box.plot plot
box.plot(x = x, names = names)
box.plot(x = x, names = names, colors = colors,
x.lab = x.lab, y.lab = y.lab, title = title, legend = legend)
box.plot(x = x, names = names, colors = colors[1])

}
\author{
David Jobst
}
