% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R
\name{bs}
\alias{bs}
\title{Brier Score}
\usage{
bs(y, t, u, na.action = na.omit, aggregate = FALSE, ...)
}
\arguments{
\item{y}{vector of observations}

\item{t}{vector of threshold values (depending on \code{y}; see details)}

\item{u}{vector of PIT values in [0,1] (depending on \code{t}; see details)}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}

\item{aggregate}{logical or function for aggregating the single scores, e.g. \code{sum}, \code{mean}, \code{weighted.mean}, ....
Default: \code{FALSE}, i.e. no aggregation function.}

\item{...}{further arguments passed to the \code{aggregate} function.}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Brier Score (BS) given observations of an univariate variable, thresholds and corresponding PIT values of a predictive distribution.
}
\details{
For a vector \code{y} of length n, the i-th entry of \code{y} belongs to the i-th entry
of \code{u} and \code{t}. For given thresholds \code{t} the PIT values are obtained by
\code{u}=F(\code{t}) for a predictive distribution F.

A lower BS indicates a better forecast.
}
\examples{
# simulated data
n <- 30
y <- rnorm(n)
t <- sample(-10:35, size = n, replace = TRUE)
u <- pnorm(t)

# bs calculation
bs(y = y, t = t, u = u)
bs(y = y, t = t, u = u, aggregate = mean)

}
\references{
Brier, G. (1950). Verification of forecasts expressed in terms of probability. Monthly Weather Review, 78, 1-3.
}
\author{
David Jobst
}
