% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.plot.R
\name{cov.plot}
\alias{cov.plot}
\title{Central Prediction Interval Coverage Model Comparison}
\usage{
cov.plot(
  x,
  models,
  interval.range,
  colors = NULL,
  title = "Interval Coverage",
  legend = "right"
)
}
\arguments{
\item{x}{matrix of central prediction interval coverages of different models in \%}

\item{models}{character; column names of \code{x} (depending on \code{x}; see details)}

\item{interval.range}{numeric; integer(s) between 0 and 100 (see details)}

\item{colors}{character; color for each object in \code{models}; default: ggplot2 setting (depending on \code{x}; see details)}

\item{title}{character; title of the plot; default: "\code{Interval Coverage}"}

\item{legend}{character; position of the legend; default: "\code{right}" (see details)}
}
\value{
ggplot object with line plots of the central prediction interval coverages of different models.
}
\description{
This function plots the central prediction interval coverage for different interval ranges of selected models as line plot.
}
\details{
For matrix \code{x}, each column contains the central prediction interval coverages of a different model in \%. Consequently
the i-th row entries of \code{x} correspond to the i-th entry in \code{interval.range}. In addition each entry in \code{models} and \code{colors} refers to
the respective column of \code{x}.
Only finite values of \code{x} and \code{interval.range} are used.

For the parameter \code{legend} the following options are available: "\code{right}", "\code{left}", "\code{bottom}", "\code{top}", "\code{hide}".

The closer the points/a line plot to the dashed line is, the more calibrated the model is.
}
\examples{
#simulated data
n <- 30
x <- matrix(runif(n)*100, ncol = 3)
x <- apply(x, 2, sort)
models <- c("A", "B", "C")
interval.range <- seq(5, 95, length.out = 10)
colors <- c("darkred", "steelblue", "orange")
title <- "Interval Coverage Comparison"
legend <- "left"

#cov.plot plot
cov.plot(x = x, models = models, interval.range = interval.range)
cov.plot(x = x, models = models, interval.range = interval.range,
colors = colors, title = title, legend = legend)

}
\author{
David Jobst
}
