% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi.R
\name{cpi}
\alias{cpi}
\title{Central Prediction Interval}
\usage{
cpi(y, lower, upper, interval.range, separate = "is", mean = FALSE)
}
\arguments{
\item{y}{vector of observations}

\item{lower}{vector with the lower quantiles (depending on \code{interval.range}; see details)}

\item{upper}{vector with the upper quantiles (depending on \code{interval.range}; see details)}

\item{interval.range}{numeric; integer(s) between 0 and 100 (see details)}

\item{separate}{character; vector with "\code{is}", "\code{overprediction}", "\code{underprediction}", "\code{width}" or "\code{coverage}"; default: "\code{is}" (see details)}

\item{mean}{logical; if \code{TRUE} the mean of the values specified in \code{separate} is calculated for output; if \code{FALSE} the single values specified in \code{separate} are used as output; default: \code{FALSE}}
}
\value{
Vector or list of score value(s) specified in \code{separate}.
}
\description{
This function can calculate the central prediction interval score (IS), overprediction, underprediction, width and coverage of a central prediction interval given observations of a one-dimensional variable and quantiles of ensemble forecasts/a predictive distribution for a certain interval range.
}
\details{
For a vector \code{y} of length n, \code{lower}, \code{upper} and \code{interval.range} should be given as vector
of length n, where the i-th entry of \code{y} belongs to the i-th entry of the other vectors. For \code{interval.range}
only one value is also sufficient. This value will be taken for all quantiles as \code{interval.range}.
Only finite values of \code{y}, \code{lower}, \code{upper} and \code{interval.range} are used.

The \code{interval.range} indicates the size of the central prediction interval, i.e. \code{interval.range} corresponds to
\code{alpha := (100 - interval.range)/100}. Consequently \code{lower} must provide the alpha/2-quantiles and \code{upper} the (1-alpha/2)-quantiles of ensemble forecasts/a predictive distribution.
Internally, \code{interval.range} will be transformed to \code{alpha}.

For ensemble forecasts of size m, it is common to choose \code{interval.range := (m-1)/(m+1)}. Consequently \code{lower} must contain
in each entry the minimum value of the m forecasts and \code{upper} must contain in each entry the maximum value of the m forecasts.

The parameter \code{separate} allows to choose the output values. "\code{is}" is the central prediction interval score, "\code{overprediction}" is the overprediction,
"\code{underprediction}" is the underprediction, "\code{width}" is the central prediction interval width and "\code{coverage}" is the central prediction interval coverage.
If an observation lies in a central prediction interval, the output will be 1, otherwise 0 for the parameter "\code{coverage}".

A lower IS indicates a better forecast.

A lower central prediction interval width score indicates a sharper forecast. The optimal value is 0.

A score of the central prediction interval coverage close to \code{interval.range} indicates a more calibrated forecast.
A central prediction interval coverage score of \code{interval.range} is optimal and indicates a calibrated forecast.
}
\examples{
#simulated data
n <- 30
y <- rnorm(n, mean = 1:n)
interval.range <- 90
alpha <- (100-interval.range)/100
lower <- qnorm(alpha/2, rnorm(n, mean = 1:n))
upper <- qnorm((1-alpha/2), rnorm(n, mean = 1:n))

#cpi calculation
cpi(y = y, lower = lower, upper = upper, interval.range = interval.range,
separate = "is", mean = FALSE)
cpi(y = y, lower = lower, upper = upper, interval.range = interval.range,
separate = "is", mean = TRUE)

cpi(y = y, lower = lower, upper = upper, interval.range = interval.range,
separate = c("is", "overprediction", "underprediction", "width", "coverage"),
mean = FALSE)
cpi(y = y, lower = lower, upper = upper, interval.range = interval.range,
separate = c("is", "overprediction", "underprediction", "width", "coverage"),
mean = TRUE)

}
\references{
Gneiting, T. and Raftery, A. (2007). Strictly Proper Scoring Rules, Prediction, and Estimation. Journal of the American Statistical Association, 102, 359-378.
}
\author{
David Jobst
}
