% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.test.R
\name{dm.test}
\alias{dm.test}
\title{Diebold-Mariano-Test}
\usage{
dm.test(
  s1,
  s2,
  alternative = c("two.sided", "less", "greater"),
  h = 1,
  na.rm = FALSE
)
}
\arguments{
\item{s1}{vector of scores from method 1}

\item{s2}{vector of scores from method 2}

\item{alternative}{character; the alternative
hypothesis and must be one of "\code{two.sided}", "\code{greater}" or
"\code{less}"; default: "\code{two.sided}"}

\item{h}{integer; forecast horizon used in calculating s1 and s2; default: 1}

\item{na.rm}{logical; if \code{TRUE} NA are stripped before the computation proceeds; if \code{FALSE} NA are used in the computation; default: \code{FALSE}

The null hypothesis is that the difference \code{s1 - s2} has zero mean.
The alternative "\code{less}" is that \code{s1 - s2} has negative mean.
The alternative "\code{greater}" is that \code{s1 - s2} has positive mean.
The alternative "\code{two.sided}" is that \code{s1 - s2} has mean unequal
zero.

The difference \code{s1 - s2} may contain missing values,
in which case complete cases are used and a warning is given.}
}
\value{
An object of class "\code{htest}".
}
\description{
This function performs the Diebold-Mariano-Test for equal predictive performance of two forecasts with respect to a scoring rule.
This test is a simplified version of \code{\link[forecast]{dm.test}}.
}
\note{
The function \code{dm.test} is inspired by the function \code{dm.test} from the R-package \code{forecast} by Hyndman et al.
}
\examples{
#simulated data
s1 <- arima.sim(list(ar = 0.7), sd = 0.5, 100)
s2 <- arima.sim(list(ar = 0.7), sd = 0.5, 100) - 0.2

#Diebold-Mariano-Test
dm.test(s1, s2)


}
\references{
Diebold, F. and Mariano, R. (1995). Comparing predictive accuracy. Journal of Business & Economic Statistics, 13, 253-263.

Gneiting, T. and Katzfuss, M. (2014). Probabilistic forecasting. Annual Review of Statistics and Its Application, 1, 125-151.
}
\author{
David Jobst
}
