% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Determinant Sharpness}
\usage{
ds(x, mean = FALSE)
}
\arguments{
\item{x}{3-dimensional array of covariance matrices of ensemble forecasts/a predictive distribution (see details)}

\item{mean}{logical; if \code{TRUE} the mean of the DS values is calculated for output; if \code{FALSE} the single DS values are used as output; default: \code{FALSE}}
}
\value{
Vector of the score value(s).
}
\description{
This function calculates the Determinant Sharpness (DS) given covariance matrices of ensemble forecasts/a predictive distribution.
}
\details{
Each third dimension entry of the array \code{x} must contain a (symmetric) covariance matrix of ensemble forecasts/a predictive distribution.
This covariance matrices are either analytically available or must be estimated by e.g. the sample covariance matrix.
Only covariance matrices with finite values in \code{x} are used.

A lower DS indicates a sharper forecast.
}
\examples{
#simulated data
n <- 30
m <- 50
x <- array(NA, dim = c(2, 2, n))
for (i in 1:n) {
x[, , i] <- cov(cbind(rnorm(m), rgamma(m, shape = 1)))
}

#ds calculation
ds(x = x, mean = FALSE)
ds(x = x, mean = TRUE)

}
\references{
Gneiting, T., Stanberry, L., Grimit, E., Held, L. and Johnson, N. (2008). Assessing probabilistic forecasts of multivariate quantities, with an application to ensemble predictions of surface winds. Test, 17, 211-264.
}
\author{
David Jobst
}
