% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dss.R
\name{dss}
\alias{dss}
\title{Dawid-Sebastiani Score}
\usage{
dss(y, x, mu = NULL, mean = FALSE)
}
\arguments{
\item{y}{vector of observations}

\item{x}{matrix of ensemble forecasts/samples of a predictive distribution or vector of variances of ensemble forecasts/a predictive distribution (depending on \code{y}; see details)}

\item{mu}{if \code{NULL}, \code{mu} is calculated by the row-wise mean of matrix \code{x}; otherwise \code{mu} must be provided as the means of the ensemble forecasts/predictive distribution; default: \code{NULL} (depending on \code{x}; see details)}

\item{mean}{logical; if \code{TRUE} the mean of the DSS values is calculated for output; if \code{FALSE} the single DSS values are used as output; default: \code{FALSE}}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Dawid-Sebastiani Score (DSS) given observations of a one-dimensional variable and ensemble forecasts/samples of a predictive distribution/parameters.
}
\details{
For a vector \code{y} of length n, \code{x} can be given as matrix of ensemble forecasts/samples of a predictive distribution
with n rows, where the i-th entry of \code{y} belongs to the i-th row
of \code{x}. The columns of \code{x} represent the samples of a predictive distribution
or ensemble forecasts. Consequently \code{mu} must be \code{NULL}. The row-wise
means and variances are determined by its sample versions.

If the variances and means of ensemble forecasts or of a predictive distribution are directly
available, \code{x} can be given as vector of variances and \code{mu} can be given as vector of means, where
the i-th entry of \code{y} belongs to the i-th entry of \code{x} and \code{mu}.
Only finite values of \code{y}, \code{x} and \code{mu} are used.

A lower DSS indicates a better forecast.
}
\examples{
#simulated data
n <- 30
m <- 50
y <- rnorm(n)
x1 <- matrix(rnorm(n*m), ncol = m)
x2 <- rep(1, n)
mu <- rep(0, n)

#dss calculation
dss(y = y, x = x1, mu = NULL, mean = FALSE)
dss(y = y, x = x1, mu = NULL, mean = TRUE)

dss(y = y, x = x2, mu = mu, mean = FALSE)
dss(y = y, x = x2, mu = mu, mean = TRUE)

}
\references{
Dawid, A. and Sebastiani, P. (1999). Coherent disperion criteria for optimal experimental design. Annals of Statistics, 27, 65-81.
}
\author{
David Jobst
}
