% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppverification-package.R
\docType{package}
\name{eppverification-package}
\alias{eppverification}
\alias{eppverification-package}
\title{eppverification: Verification Tools for the Statistical Postprocessing of Ensemble Forecasts}
\description{
This package provides user-friendly univariate and multivariate verification tools for the statistical ensemble post-processing. It allows to score and assess the calibration (reliability) and sharpness of ensemble forecasts and predictive distributions. In addition this package can be used to create useful contemporary visualizations for verification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jobstdavid/eppverification}
  \item Report bugs at \url{https://github.com/jobstdavid/eppverification/issues}
}

}
\author{
\strong{Maintainer}: David Jobst \email{jobstd@uni-hildesheim.de} (10.5281/zenodo.5574572)

}
\keyword{internal}
