% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.plot.R
\name{line.plot}
\alias{line.plot}
\title{Line Plot}
\usage{
line.plot(
  x,
  y,
  names,
  point = FALSE,
  linetypes = "solid",
  colors = "black",
  x.lab = "",
  y.lab = "",
  title = "",
  legend = "right"
)
}
\arguments{
\item{x}{evaluation points of class "\code{Date}", "\code{numeric}" or "\code{integer}"}

\item{y}{matrix of forecast data (see details)}

\item{names}{character; column names of \code{y} (depending on \code{y}; see details)}

\item{point}{logical; if \code{TRUE} points are additionally plotted for each object in \code{names}; if \code{FALSE} no points are plotted for each object in \code{names}; default: \code{FALSE}}

\item{linetypes}{character; line type for each object in \code{names}; default: "\code{solid}" (depending on \code{y}; see details)}

\item{colors}{character; color for each object in \code{names}; default: "\code{black}" (depending on \code{y}; see details)}

\item{x.lab}{character; label of x-axis; default: ""}

\item{y.lab}{character; label of y-axis; default: ""}

\item{title}{character; title of the plot; default: ""}

\item{legend}{character; position of the legend; default: "\code{right}" (see details)}
}
\value{
ggplot object with line plots of the forecast data.
}
\description{
This function plots forecast data as line plots.
}
\details{
For \code{x} of length n, \code{y} should be given as matrix with n rows, where each column consists of one forecast.
The the i-th entry of \code{x} belongs to the i-th row of \code{y}. In addition each entry in \code{names}, \code{linetypes} and \code{colors} refers to
the respective column of \code{y}. If only one value is provided for \code{names}, \code{linetypes} and \code{colors} this value is
used for all column entries in \code{y}.
Only finite values of \code{x} and \code{y} are used.

For \code{linetypes} the specifications of \code{linetype} in ggplot2 are used.

For \code{colors} the specifications of \code{color} in ggplot2 are used.

For the parameter \code{legend} the following options are available: "\code{right}", "\code{left}", "\code{bottom}", "\code{top}", "\code{hide}".
}
\examples{
#simulated data
n <- 30
k <- 10
x1 <- 1:n
x2 <- seq(Sys.Date(), by = "day", length.out = n)
obs <- rnorm(n, mean = 1:n)
y <- sapply(1:k, function(i) obs-(i-1)*runif(n))
names1 <- c("observation", paste("forecast", sep = " ", 1:(k-1)))
names2 <- c("observation", rep("forecasts", 7), rep("modified forecasts", 2))
linetypes1 <- c("solid", rep("dashed", k-1))
linetypes2 <- c("solid", rep("dashed", 7), rep("dotted", 2))
colors1 <- c("red", rep("black", k-1))
colors2 <- c("red", rep("black", 7), rep("blue", 2))
x.lab <- "Date"
y.lab <- "Value"
title <- "Observation vs. Forecast"
legend1 <- "hide"
legend2 <- "bottom"


#line.plot plot
line.plot(x = x1, y = y, names = names1)
line.plot(x = x1, y = y, names = names1, point = TRUE,
linetypes = linetypes1, legend = legend1)
line.plot(x = x2, y = y, names = names1, colors = colors1,
x.lab = x.lab, y.lab = y.lab, title = title, legend = legend2)
line.plot(x = x2, y = y, names = names1, colors = colors1[1],
linetypes = linetypes1[2])
line.plot(x = x2, y = y, names = names2, colors = colors2,
linetypes = linetypes2)

}
\author{
David Jobst
}
