% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{logs}
\alias{logs}
\title{Logarithmic Score}
\usage{
logs(y, na.action = na.omit, aggregate = FALSE, ...)
}
\arguments{
\item{y}{vector of density values}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}

\item{aggregate}{logical or function for aggregating the single scores, e.g. \code{sum}, \code{mean}, \code{weighted.mean}, ....
Default: \code{FALSE}, i.e. no aggregation function.}

\item{...}{further arguments passed to the \code{aggregate} function.}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Logarithmic Score (LogS) given density values of observations of an uni- or multivariate predictive distribution.
}
\details{
For a predictive density function f the density values \code{y} of observations \code{x} are
obtained by \code{y} = f(\code{x}).

A lower LogS indicates a better forecast.
}
\examples{
# simulated data
n <- 30
x <- sample(-10:35, size = n, replace = TRUE)
y <- dnorm(x = x, mean = 0, sd = 1)

# logs calculation
logs(y = y)
logs(y = y, aggregate = mean)

}
\references{
Good, I. (1952). Rational decisions. Journal of the Royal Statistical Society Ser. B, 14, 107-114.
}
\author{
David Jobst
}
