% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.pit.R
\name{m.pit}
\alias{m.pit}
\title{Expectation of the Probability Integral Transform Values}
\usage{
m.pit(u)
}
\arguments{
\item{u}{vector of PIT values in [0,1] (see details)}
}
\value{
Expectation value of the PIT values.
}
\description{
This function calculates the Expectation value of the Probability Integral Transform values (E(PIT)).
}
\details{
The vector \code{u} contains the PIT values \code{u}=F(\code{x}) for a predictive
distribution F and argument \code{x}.
The expectation value is calculated by the sample mean of the PIT values.

The expectation value of the PIT values (E(PIT)) provides information on the bias of a predictive distribution.
An expectation of the PIT values equal to 1/2 corresponds to the expectation of the uniform distribution
on [0,1], which is desirable. Any deviation from 1/2 indicates that the predictive distribution is biased.
}
\examples{
#simulated data
n <- 10000
u <- runif(n)

#E(PIT) calculcation
m.pit(u = u)

}
\references{
Gneiting, T. and Ranjan, R. (2013). Combining predictive distributions. Electronic Journal of Statistics, 7, 1747-1782.
}
\author{
David Jobst
}
