% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.plot.R
\name{map.plot}
\alias{map.plot}
\title{Map Plot}
\usage{
map.plot(
  coordinates,
  names,
  maptype = "terrain",
  size = 2,
  shape = 16,
  colors = "black",
  x.lab = "Longitude",
  y.lab = "Latitude",
  title = "",
  legend = "right"
)
}
\arguments{
\item{coordinates}{matrix of longitude and latitude coordinates (see details)}

\item{names}{character; row names of \code{coordinates} (depending on \code{coordinates}; see details)}

\item{maptype}{character; type of map; default: "\code{terrain}" (see details)}

\item{size}{numeric; integer specifying the size for each object in \code{names}; default: \code{2}}

\item{shape}{numeric; integer specifying the shape for each object in \code{names}; default: \code{16} (depending on \code{coordinates}; see details)}

\item{colors}{character; color for each object in \code{names}; default: "\code{black}" (depending on \code{coordinates}; see details)}

\item{x.lab}{character; label of x-axis; default: "\code{Longitude}"}

\item{y.lab}{character; label of y-axis; default: "\code{Latitude}"}

\item{title}{character; title of the plot; default: ""}

\item{legend}{character; position of the legend; default: "\code{right}" (see details)}
}
\value{
ggplot object with a map plot for given coordinates.
}
\description{
This function creates map plots for given coordinates.
}
\details{
In \code{coordinates} the first columns stands for the longitude and the
second column for the latitude coordinate. Furthermore the i-th row entry of \code{coordinates} corresponds to the i-th entry
in \code{names}, \code{size}, \code{shape} and \code{colors}. If only one value is provided for \code{names}, \code{size}, \code{shape} and \code{colors} this value is
used for all row entries in \code{coordinates}.
Only finite values of \code{coordinates} are used.

As map source "\code{stamen}" is chosen. Consequently the available maptypes are: "\code{terrain}",
"\code{terrain-background}", "\code{terrain-labels}", "\code{terrain-lines}", "\code{toner}",
"\code{toner-2010}", "\code{toner-2011}", "\code{toner-background}", "\code{toner-hybrid}",
"\code{toner-labels}", "\code{toner-lines}", "\code{toner-lite}", "\code{watercolor}".

For the parameters \code{size}, \code{shape} and \code{colors} the specifications of ggplot2 are available.

For the parameter \code{legend} the following options are available: "\code{right}", "\code{left}", "\code{bottom}", "\code{top}", "\code{hide}".
}
\examples{
#simulated data
coordinates <- c(13.38, 52.53, 2.35, 48.86, -0.13, 51.51, -3.70, 40.42)
coordinates <- matrix(coordinates, ncol = 2, byrow = TRUE)
names1 <- c("Berlin", "Paris", "London", "Madrid")
names2 <- c("stations")
maptype1 <- "toner-lite"
maptype2 <- "terrain-background"
size1 <- c(1, 3, 5, 7)
size2 <- 5
shape1 <- 16:19
shape2 <- 17
colors1 <- c("darkred", "steelblue", "orange", "darkgreen")
colors2 <- "darkred"
x.lab <- "Lon"
y.lab <- "Lat"
title <- "Map"
legend <- "bottom"

#map.plot plot
map.plot(coordinates = coordinates, names = names1)
map.plot(coordinates = coordinates, names = names1, maptype = maptype1,
size = size1, shape = shape1, colors = colors1, x.lab = x.lab, y.lab = y.lab,
title = title, legend = legend)
map.plot(coordinates = coordinates, names = names2, maptype = maptype2,
size = size2, shape = shape2, colors = colors2)

}
\author{
David Jobst
}
