% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrnk.R
\name{mrnk}
\alias{mrnk}
\title{Multivariate Ranks}
\usage{
mrnk(y, x, method = "mv", na.rm = FALSE)
}
\arguments{
\item{y}{matrix of observations (see details)}

\item{x}{3-dimensional array of samples of a predictive distribution (depending on \code{y}; see details)}

\item{method}{character; "\code{mv}", "\code{avg}", "\code{mst}", "\code{bd}"; default: "\code{mv}" (see details)}

\item{na.rm}{logical; if \code{TRUE} NA are stripped before the rank computation proceeds; if \code{FALSE} NA are used in the rank computation; default: \code{FALSE}}
}
\value{
Vector of ranks.
}
\description{
This function calculates the ranks given observations of a multivariate variable and samples of a predictive distribution.
}
\details{
The observations are given in the matrix \code{y} with n rows, where each column belongs to an univariate observation variable.
The i-th row of matrix \code{y} belongs to the i-th third dimension entry of the array \code{x}. The i-th third dimension
entry must be a matrix with n rows, having the same structure as \code{y}, filled with the samples of a multivariate predictive distribution.

For the calculation of the ranks, different methods are available, where "\code{mv}" stands for "multivariate ranks",
"\code{avg}" stands for "average ranks", "\code{mst}" stands for "minimum-spanning-tree ranks" and
"\code{bd}" stands for "band-depth ranks". These methods are implemented as described in e.g. Thorarinsdottir et al. (2016).
}
\examples{
# simulated data
n <- 30
m <- 50
y <- cbind(rnorm(n), rgamma(n, shape = 1))
x <- array(NA, dim = c(m, 2, n))
x[, 1, ] <- rnorm(n*m)
x[, 2, ] <- rgamma(n*m, shape = 1)

# mrnk calculation
mrnk(y = y, x = x, method = "mv")
mrnk(y = y, x = x, method = "avg")
mrnk(y = y, x = x, method = "mst")
mrnk(y = y, x = x, method = "bd")

}
\references{
Gneiting, T., Stanberry, L., Grimit, E., Held, L. and Johnson, N. (2008). Assessing probabilistic forecasts of multivariate quantities, with an application to ensemble predictions of surface winds. Test, 17, 211-264.

Smith, L. and Hansen, J. (2004). Extending the limits of ensemble forecast verification with the minimum spanning tree. Monthly Weather Review, 132, 1522-1528.

Thorarinsdottir, T., Scheurer, M. and Heinz, C. (2016). Assessing the calibration of high-dimensional ensemble forecasts using rank histograms. Journal of Computational and Graphical Statistics, 25, 105-122.

Wilks, D. (2004). The minimum spanning tree histogram as verification tool for multidimensional ensemble forecasts. Monthly Weather Review, 132, 1329-1340.
}
\author{
David Jobst
}
