% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{qs}
\alias{qs}
\title{Quantile Score}
\usage{
qs(y, q, p, na.action = na.omit, aggregate = FALSE, ...)
}
\arguments{
\item{y}{vector of observations}

\item{q}{vector of quantile values (depending on \code{p}; see details)}

\item{p}{vector of probabilities in (0,1)}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}

\item{aggregate}{logical or function for aggregating the single scores, e.g. \code{sum}, \code{mean}, \code{weighted.mean}, ....
Default: \code{FALSE}, i.e. no aggregation function.}

\item{...}{further arguments passed to the \code{aggregate} function.}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Quantile Score (QS) given observations of an univariate variable, probabilities and corresponding quantiles of a predictive distribution.
}
\details{
For a vector \code{y} of length n, the i-th entry of \code{y} belongs to the i-th entry
of \code{q} and \code{p}. For given probabilities \code{p} the quantile values are obtained by
\code{q}=F^{-1}(\code{p}) for a predictive distribution F.

A lower QS indicates a better forecast.
}
\examples{
# simulated data
n <- 30
y <- rnorm(n)
p <- runif(n)
q <- qnorm(p)

# qs calculation
qs(y = y, q = q, p = p)
qs(y = y, q = q, p = p, aggregate = mean)

}
\references{
Gneiting, T. (2011). Making and Evaluating Point Forecasts. Journal of the American Statistical Association, 106(494), 746-762.
}
\author{
David Jobst
}
