% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri.R
\name{ri}
\alias{ri}
\title{Reliability Index}
\usage{
ri(y, x, bins = NULL)
}
\arguments{
\item{y}{vector of observations}

\item{x}{matrix of ensemble forecasts/samples of a predictive distribution (depending on \code{y}; see details)}

\item{bins}{numeric; if \code{NULL} the number of bins is equal to \code{ncol(x)+1}; otherwise \code{bins} must be chosen so that \code{(ncol(x)+1)/bins} is an integer; default: \code{NULL} (see details)}
}
\value{
Vector of the score value.
}
\description{
This function calculates the Reliability Index (RI) given observations of a one-dimensional variable and ensemble forecasts/samples of a predictive distribution.
}
\details{
For a vector \code{y} of length n, \code{x} should be given as matrix
with n rows, where the i-th entry of \code{y} belongs to the i-th row
of \code{x}. The columns of \code{x} represent the samples of a predictive distribution
or ensemble forecasts.
Only finite values of \code{y} and \code{x} are used.

The parameter \code{bins} specifies the number of columns for the VRH. For "large"
\code{ncol(x)} it is often reasonable to reduce the resolution of the VRH by
using \code{bins} so that \code{(ncol(x)+1)/bins} is an integer.

The deviation from uniformity of the VRH can be quantified by the reliability index (RI).
The smaller the RI, the better is the calibration of the forecast. The
optimal value of the RI is 0.
}
\examples{
#simulated data
n <- 30
m <- 50
y <- rnorm(n)
x <- matrix(rnorm(n*m), ncol = m)

#ri calculation
ri(y = y, x = x)
ri(y = y, x = x, bins = 17)

}
\references{
Delle Monache, L., Hacker, J., Zhou, Y., Deng, X. and Stull, R., (2006). Probabilistic aspects of meteorological and ozone regional ensemble forecasts. Journal of Geophysical Research: Atmospheres, 111, D24307.
}
\author{
David Jobst
}
