% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.plot.R
\name{score.plot}
\alias{score.plot}
\title{Score Plot}
\usage{
score.plot(x, models, measures, title = "Model Comparison")
}
\arguments{
\item{x}{matrix of scores (see detail)}

\item{models}{character; column names of \code{x} (depending on \code{x}; see details)}

\item{measures}{character; row names of \code{x} (depending on \code{x}; see details)}

\item{title}{character; title of the plot; default: "\code{Model Comparison}"}
}
\value{
ggplot object with score plot of different models rated by selected measures.
}
\description{
This function plots scores of different models rated by selected measures as heatmap.
}
\details{
For matrix \code{x}, each row contains the scores of one measure and
each columns contains the scores belonging to one model.
Only finite values of \code{x} are used.

The brighter the color in the plot, the lower the score is.
}
\examples{
#simulated data
x <- matrix(c(0.5, 0.3, 0.8, 0.21, 1.5, 0.7, 2, 1), byrow = TRUE, ncol = 4)
models <- c("A", "B", "C", "D")
measures <- c("CRPS", "LogS")
title <- ""

#score.plot plot
score.plot(x = x, models = models, measures = measures)
score.plot(x = x, models = models, measures = measures, title = title)

}
\author{
David Jobst
}
