% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{ss}
\alias{ss}
\title{Skill Score}
\usage{
ss(s1, s2, na.action = na.omit, aggregate = FALSE, ...)
}
\arguments{
\item{s1}{vector of mean scores from method 1}

\item{s2}{vector of mean scores from method 2}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}

\item{aggregate}{logical or function for aggregating the single scores, e.g. \code{sum}, \code{mean}, \code{weighted.mean}, ....
Default: \code{FALSE}, i.e. no aggregation function.}

\item{...}{further arguments passed to the \code{aggregate} function.}
}
\value{
Skill score.
}
\description{
This function calculates the Skill Score (SS) of the mean score time series.
}
\details{
The Skill Score (SS) uses the mean scores \code{s1} of the forecasting method of interest (method 1) and \code{s2} of the benchmark method (method 2).
Consequently, the Skill Score can be calculated via

\deqn{\text{SS}=1-\frac{s_1}{s_2},}

indicating the relative improvement of method 1 over method 2.

Skill scores are positively oriented with a maximum value of 1.
Positive values indicate an improvement of method 1 over method 2, negative values stand for a worse performance of method 1 over method 2.
}
\examples{
# simulated data
s1 <- rnorm(100)
s2 <- rnorm(100)

ss(s1 = s1, s2 = s2)
ss(s1 = s1, s2 = s2, aggregate = mean)

}
\references{
Gneiting, T. and Raftery, A. (2007). Strictly proper scoring rules, prediction, and estimation. Journal of the American Statistical Association, 102, 359-378.
}
\author{
David Jobst
}
