% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.pit.R
\name{var.pit}
\alias{var.pit}
\title{Variance of the Probability Integral Transform Values}
\usage{
var.pit(u)
}
\arguments{
\item{u}{vector of PIT values in [0,1] (see details)}
}
\value{
Variance of the PIT values.
}
\description{
This function calculates the variance of the Probability Integral Transform values (Var(PIT)).
}
\details{
The vector \code{u} contains the PIT values \code{u}=F(\code{x}) for a predictive
distribution F and argument \code{x}.
Only finite values of \code{u} are used.

The variance of the PIT values (Var(PIT)) provides information on the dispersion of a predictive distribution.
A variance of the PIT values equal to 1/12 ≈ 0.0833 corresponds to the variance of the uniform distribution
on [0,1], which is desirable. A variance greater than 1/12 indicates underdispersion and a
variance smaller than 1/12 indicates overdispersion of the predictive distribution.
}
\examples{
#simulated data
n <- 10000
u <- runif(n)

#Var(PIT) calculcation
var.pit(u = u)

}
\references{
Gneiting, T. and Ranjan, R. (2013). Combining predictive distributions. Electronic Journal of Statistics, 7, 1747-1782.
}
\author{
David Jobst
}
