% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr.hist.R
\name{vr.hist}
\alias{vr.hist}
\title{Verification Rank Histogram}
\usage{
vr.hist(
  y,
  x,
  bins = NULL,
  type = "relative",
  title = "Verification Rank Histogram",
  ri = FALSE,
  ent = FALSE
)
}
\arguments{
\item{y}{vector of observations}

\item{x}{matrix of ensemble forecasts/samples of a predictive distribution (depending on \code{y}; see details)}

\item{bins}{numeric; if \code{NULL} the number of bins is equal to \code{ncol(x)+1}; otherwise \code{bins} must be chosen so that \code{(ncol(x)+1)/bins} is an integer; default: \code{NULL} (see details)}

\item{type}{character; "\code{relative}", "\code{absolute}" and "\code{density}"; default: "\code{relative}" (see details)}

\item{title}{character; title of the plot; default: "\code{Verification Rank Histogram}"}

\item{ri}{logical; if \code{TRUE} the reliability index is calculated for the plot (see details); if \code{FALSE} the reliability index is not calculated; default: \code{FALSE}}

\item{ent}{logical; if \code{TRUE} the entropy is calculated for the plot (see details); if \code{FALSE} the entropy  is not calculated; default: \code{FALSE}}
}
\value{
ggplot object with a plot of the Verification Rank Histogram.
}
\description{
This function plots the Verification Rank Histogram (VRH) given observations of a one-dimensional variable and ensemble forecasts/samples of a predictive distribution.
}
\details{
For a vector \code{y} of length n, \code{x} should be given as matrix
with n rows, where the i-th entry of \code{y} belongs to the i-th row
of \code{x}. The columns of \code{x} represent the samples of a predictive distribution
or ensemble forecasts.
Only finite values of \code{y} and \code{x} are used.

The parameter \code{bins} specifies the number of columns for the VRH. For "large"
\code{ncol(x)} it is often reasonable to reduce the resolution of the VRH by
using \code{bins} so that \code{(ncol(x)+1)/bins} is an integer.

If \code{type} is "\code{relative}" the relative frequencies of the bins are plotted.
If \code{type} is "\code{absolute}" the absolute frequencies of the bins are plotted.
If \code{type} is "\code{density}" the relative densities of the bins are plotted.

An uniform VRH indicates a calibrated predictive distribution or ensemble forecasts. A ∩-shape in the
VRH indicates overdispersion and a ∪-shape indicates underdispersion
of the predictive distribution or ensemble forecasts. A systematic bias of the predictive distribution or ensemble forecasts
results in a triangular shaped VRH histogram.

The deviation from uniformity of the VRH can be quantified by the reliability index (RI).
The smaller the RI, the better is the calibration of the forecast. The
optimal value of the RI is 0.

The entropy is a tool to assess the calibration of a forecast. The optimal
value of the entropy is 1, representing a calibrated forecast.
}
\examples{
#simulated data
n <- 30
m <- 50
y <- rnorm(n)
x <- matrix(rnorm(n*m), ncol = m)

#vr.hist plot
vr.hist(y = y, x = x)
vr.hist(y = y, x = x, bins = 17, title = "VRH", ri = TRUE, ent = FALSE)
vr.hist(y = y, x = x, bins = 3, type = "absolute", ri = FALSE, ent = TRUE)
vr.hist(y = y, x = x, bins = 3, type = "density", ri = TRUE, ent = TRUE)

}
\references{
Anderson, J. (1996). A method for producing and evaluating probabilistic forecasts from ensemble model integrations. Journal of Climate, 9, 1518-1530.

Candille, G. and Talagrand, O. (2005). Evaluation of probabilistic prediction systems for a scalar variable. Quarterly Journal of the Royal Meteorological Society, 131(609), 2131-2150.

Delle Monache, L., Hacker, J., Zhou, Y., Deng, X. and Stull, R., (2006). Probabilistic aspects of meteorological and ozone regional ensemble forecasts. Journal of Geophysical Research: Atmospheres, 111, D24307.

Hamill, T. and Colucci, S. (1997). Verification of Eta-RSM short-range ensemble forecasts. Monthly Weather Review, 125, 1312-1327.

Hamill, T. (2001). Interpretation of rank histograms for verifying ensemble forecasts. Monthly Weather Review, 129, 550-560.

Taillardat, M., Mestre, O., Zamo, M. and Naveau, P., (2016). Calibrated Ensemble Forecasts Using Quantile Regression Forests and Ensemble Model Output Statistics. American Meteorological Society, 144, 2375-2393.

Tribus, M. (1969). Rational Descriptions, Descisions and Designs. Pergamon Press.

Talagrand, O., Vautard, R. and Strauss, B. (1997). Evaluation of probabilistic prediction systems. Workshop on Predictability (ECMWF), 1-25.
}
\author{
David Jobst
}
