% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{ae}
\alias{ae}
\title{Absolute Error}
\usage{
ae(y, x, mean = FALSE)
}
\arguments{
\item{y}{vector of observations}

\item{x}{matrix of ensemble forecasts/samples of a predictive distribution or vector of medians of ensemble forecasts/a predictive distribution (depending on \code{y}; see details)}

\item{mean}{logical; if \code{TRUE} the mean of the AE values is calculated for output; if \code{FALSE} the single AE values are used as output; default: \code{FALSE}}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Absolute Error (AE) given observations of a one-dimensional variable and ensemble forecasts/samples of a predictive distribution.
}
\details{
For a vector \code{y} of length n, \code{x} can be given as matrix of ensemble forecasts/samples of a predictive distribution
with n rows, where the i-th entry of \code{y} belongs to the i-th row
of \code{x}. The columns of \code{x} represent the samples of a predictive distribution
or ensemble forecasts. The row-wise medians are determined by its sample version.

If the medians of ensemble forecasts or of a predictive distribution are directly
available, \code{x} can be given as vector of medians, where
the i-th entry of \code{y} belongs to the i-th entry of \code{x}.
Only finite values of \code{y} and \code{x} are used.

In addition, with this function, the Mean Absolute Error (MAE) can be calculated.

A lower AE indicates a better forecast.
}
\examples{
#simulated data
n <- 30
m <- 50
y <- rnorm(n)
x1 <- matrix(rnorm(n*m), ncol = m)
x2 <- apply(x1, 1, median)

#ae calculation
ae(y = y, x = x1, mean = FALSE)
ae(y = y, x = x1, mean = TRUE)

ae(y = y, x = x2, mean = FALSE)
(mae <- ae(y = y, x = x2, mean = TRUE))

}
\references{
Vannitsem, S., Wilks, D. and Messner, J. (2018). Statistical Postprocessing of Ensemble Forecasts. Elsevier. 164.
}
\author{
David Jobst
}
