% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhp.R
\name{bh.test}
\alias{bh.test}
\title{Benjamini-Hochberg-Procedure}
\usage{
bh.test(p, alpha = 0.05, na.rm = FALSE)
}
\arguments{
\item{p}{vector of p-values on which the Benjamini-Hochberg-Procedure should be applied}

\item{alpha}{numeric; false discovery rate at level alpha; alpha is number between 0 and 1}

\item{na.rm}{logical; if \code{TRUE} NA are removed after the computation; if \code{FALSE} NA are used in the computation; default: \code{FALSE}}
}
\value{
A logical vector where the value TRUE indicates that the p-value at the
corresponding  position is significant after the Benjamini-Hochberg-Procedure
with false discovery rate alpha. If the value is FALSE the p-value at the
corresponding  position is not significant after the
Benjamini-Hochberg-Procedure with false discovery rate alpha.
}
\description{
This function performs the Benjamini-Hochberg-Procedure for different p-values in the multiple testing setting.
This function computes the significance of the individual tests.
}
\examples{
#simulated data
p <- runif(100)

#Benjamini-Hochberg-Procedure
bh.test(p = p)
bh.test(p = p, alpha = 0.1)


}
\references{
Benjamini, Y. and Hochberg, Y. (1995). Controlling the False Discovery Rate: A Practical and Powerful Approach to Multiple Testing. Journal of the Royal Statistical Society, 57, 289-300.
}
\author{
David Jobst
}
