% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R
\name{bs}
\alias{bs}
\title{Brier Score}
\usage{
bs(y, z, p, mean = FALSE)
}
\arguments{
\item{y}{vector of observations}

\item{z}{vector of threshold values (depending on \code{y}; see details)}

\item{p}{vector of PIT values in [0,1] (depending on \code{z}; see details)}

\item{mean}{logical; if \code{TRUE} the mean of the BS values is calculated for output; if \code{FALSE} the single BS values are used as output; default: \code{FALSE}}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Brier Score (BS) given observations of a one-dimensional variable, thresholds and corresponding PIT values of a predictive distribution.
}
\details{
For a vector \code{y} of length n, the i-th entry of \code{y} belongs to the i-th entry
of \code{z} and \code{p}. For given thresholds \code{z} the PIT values are obtained by
\code{p}=F(\code{z}) for a predictive distribution F. Only finite values of \code{y}, \code{z} and \code{p} are used.

A lower BS indicates a better forecast.
}
\examples{
#simulated data
n <- 30
y <- rnorm(n)
z <- sample(-10:35, size = n, replace = TRUE)
p <- pnorm(z)

#bs calculation
bs(y = y, z = z, p = p, mean = FALSE)
bs(y = y, z = z, p = p, mean = TRUE)

}
\references{
Brier, G. (1950). Verification of forecasts expressed in terms of probability. Monthly Weather Review, 78, 1-3.
}
\author{
David Jobst
}
