% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi.plot.R
\name{cpi.plot}
\alias{cpi.plot}
\title{Central Prediction Intervals Plot}
\usage{
cpi.plot(
  x,
  y,
  lower,
  upper,
  nominal.coverage,
  x.lab = "",
  y.lab = "",
  title = paste0(nominal.coverage, "\% Central Prediction Interval"),
  info = FALSE
)
}
\arguments{
\item{x}{evaluation points of class "\code{Date}", "\code{numeric}" or "\code{integer}"}

\item{y}{vector of observations}

\item{lower}{vector with the lower quantiles (depending on \code{nominal.coverage}; see details)}

\item{upper}{vector with the upper quantiles (depending on \code{nominal.coverage}; see details)}

\item{nominal.coverage}{numeric; nominal coverage in \% (see details)}

\item{x.lab}{character; label of x-axis; default: ""}

\item{y.lab}{character; label of y-axis; default: ""}

\item{title}{character; title of the plot; default: "\code{nominal.coverage\% Central Prediction Interval}"}

\item{info}{logical; if \code{TRUE} the mean central prediction interval width and coverage are calculated for the plot (see details); if \code{FALSE} the mean central prediction interval width and coverage are not calculated; default: \code{FALSE}}
}
\value{
ggplot object with a plot of the central prediction intervals.
}
\description{
This function plots the central prediction intervals given evaluation points, observations of an univariate variable and quantiles of a predictive distribution for a certain interval range.
}
\details{
For \code{x} of length n, \code{y}, \code{lower} and \code{upper} should be given as vector
of length n, where the i-th entry of \code{x} belongs to the i-th entry of the other vectors.
NA's are omitted.

A lower central prediction interval width indicates a sharper forecast.

A score of the central prediction interval coverage close to \code{nominal.coverage} indicates a more calibrated forecast.
A central prediction interval coverage score of \code{nominal.coverage} is optimal and indicates a calibrated forecast.
}
\examples{
# simulated data
n <- 30
x1 <- 1:n
x2 <- seq(Sys.Date(), by = "day", length.out = n)
y <- rnorm(n, mean = 1:n)
nominal.coverage <- 90
alpha <- (100-nominal.coverage)/100
lower <- qnorm(alpha/2, rnorm(n, mean = 1:n))
upper <- qnorm((1-alpha/2), rnorm(n, mean = 1:n))
x.lab <- "Date"
y.lab <- "Value"
title <- "Central Prediction Interval"


# cpi.plot plot
cpi.plot(x = x1, y = y, lower = lower, upper = upper,
nominal.coverage = nominal.coverage)
cpi.plot(x = x2, y = y, lower = lower, upper = upper,
nominal.coverage = nominal.coverage, x.lab = x.lab, y.lab = y.lab,
title = title, info = TRUE)

}
\references{
Gneiting, T. and Raftery, A. (2007). Strictly Proper Scoring Rules, Prediction,and Estimation. Journal of the American Statistical Association, 102, 359-378.
}
\author{
David Jobst
}
