% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Determinant Sharpness}
\usage{
ds(x, covmat = FALSE, na.action = na.omit, aggregate = FALSE, ...)
}
\arguments{
\item{x}{3-dimensional array of samples or covariance matrices of a predictive distribution (see details)}

\item{covmat}{logical; if \code{TRUE} the covariance matrices of a predictive distribution are provided in \code{x} by the user;
if \code{FALSE} the sample covariance matrix is calculated based on the samples of a predictive distribution contained in \code{x}; default: \code{FALSE}}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}

\item{aggregate}{logical or function for aggregating the single scores, e.g. \code{sum}, \code{mean}, \code{weighted.mean}, ....
Default: \code{FALSE}, i.e. no aggregation function.}

\item{...}{further arguments passed to the \code{aggregate} function.}
}
\value{
Vector of the score value(s).
}
\description{
This function calculates the Determinant Sharpness (DS) given covariance matrices of a predictive distribution.
}
\details{
Each third dimension entry of the array \code{x} must contain a (symmetric) covariance matrix or samples of a multivariate predictive distribution.
This covariance matrices are either analytically available or must are estimated by e.g. the sample covariance matrix.

A lower DS indicates a sharper forecast.
}
\examples{
# simulated data
n <- 30
m <- 50
x <- array(NA, dim = c(2, 2, n))
for (i in 1:n) {
x[, , i] <- cov(cbind(rnorm(m), rgamma(m, shape = 1)))
}
z <- array(rnorm(10000*5*365), dim = c(10000, 5, 365))

# ds calculation with provided covariance matrix
ds(x = x, covmat = TRUE)
ds(x = x, covmat = TRUE, aggregate = mean)

# ds calculation without provided covariance matrix
ds(x = z, covmat = FALSE)
ds(x = z, covmat = FALSE, aggregate = mean)

}
\references{
Gneiting, T., Stanberry, L., Grimit, E., Held, L. and Johnson, N. (2008). Assessing probabilistic forecasts of multivariate quantities, with an application to ensemble predictions of surface winds. Test, 17, 211-264.
}
\author{
David Jobst
}
