% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ent.R
\name{ent}
\alias{ent}
\title{Entropy}
\usage{
ent(y, x, bins = NULL, na.rm = FALSE)
}
\arguments{
\item{y}{vector of observations}

\item{x}{matrix of ensemble forecasts/samples of a predictive distribution (depending on \code{y}; see details)}

\item{bins}{numeric; if \code{NULL} the number of bins is equal to \code{ncol(x)+1}; otherwise \code{bins} must be chosen so that \code{(ncol(x)+1)/bins} is an integer; default: \code{NULL} (see details)}

\item{na.rm}{logical; if \code{TRUE} NA are stripped before the computation proceeds; if \code{FALSE} NA are used in the computation; default: \code{FALSE}}
}
\value{
Vector of the score value.
}
\description{
This function calculates the Entropy given observations of a one-dimensional variable and ensemble forecasts/samples of a predictive distribution.
}
\details{
For a vector \code{y} of length n, \code{x} should be given as matrix
with n rows, where the i-th entry of \code{y} belongs to the i-th row
of \code{x}. The columns of \code{x} represent the samples of a predictive distribution
or ensemble forecasts.

The parameter \code{bins} specifies the number of columns for the VRH. For "large"
\code{ncol(x)} it is often reasonable to reduce the resolution of the VRH by
using \code{bins} so that \code{(ncol(x)+1)/bins} is an integer.

The entropy is a tool to assess the calibration of a forecast. The optimal
value of the entropy is 1, representing a calibrated forecast.
}
\examples{
#simulated data
n <- 30
m <- 50
y <- rnorm(n)
x <- matrix(rnorm(n*m), ncol = m)

#entropy calculation
ent(y = y, x = x, bins = 3)

}
\references{
Delle Monache, L., Hacker, J., Zhou, Y., Deng, X. and Stull, R., (2006). Probabilistic aspects of meteorological and ozone regional ensemble forecasts. Journal of Geophysical Research: Atmospheres, 111, D24307.

Tribus, M. (1969). Rational Descriptions, Descisions and Designs. Pergamon Press.

Taillardat, M., Mestre, O., Zamo, M. and Naveau, P., (2016). Calibrated Ensemble Forecasts Using Quantile Regression Forests and Ensemble Model Output Statistics. American Meteorological Society, 144, 2375-2393.
}
\author{
David Jobst
}
