% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{es}
\alias{es}
\title{Energy Score}
\usage{
es(y, x, method = "ens", mean = FALSE)
}
\arguments{
\item{y}{matrix of observations (see details)}

\item{x}{3-dimensional array of ensemble forecasts/samples of a predictive distribution (depending on \code{y}; see details)}

\item{method}{character; "\code{ens}" and "\code{mc}"; default: "\code{ens}" (see details)}

\item{mean}{logical; if \code{TRUE} the mean of the ES values is calculated for output; if \code{FALSE} the single ES values are used as output; default: \code{FALSE}}
}
\value{
Vector of the score value(s).
}
\description{
This function calculates the Energy Score (ES) given observations of a multivariate variable and ensemble forecasts/samples of a predictive distribution.
}
\details{
The observations are given in the matrix \code{y} with n rows, where each column belongs to an univariate observation variable.
The i-th row of matrix \code{y} belongs to the i-th third dimension entry of the array \code{x}. The i-th third dimension
entry must be a matrix with n rows, having the same structure as \code{y}, filled with the ensemble forecasts or samples of a predictive distribution.
Only finite values of \code{y} and \code{x} are used.

If method "\code{ens}" is specified, the ES values are calculated for
given ensemble forecasts in \code{x} (Gneiting et al., 2008). If method "\code{mc}"
is specified, the ES values are calculated by a Monte-Carlo approximation using samples
of a predictive distribution in \code{x} (Gneiting et al., 2008). In the latter case, the number of
samples should be "high", e.g. 10.000.

A lower ES indicates a better forecast.
}
\examples{
#simulated data
n <- 30
m1 <- 50
m2 <- 10000
y <- cbind(rnorm(n), rgamma(n, shape = 1))
x1 <- array(NA, dim = c(m1, 2, n))
x2 <- array(NA, dim = c(m2, 2, n))
x1[, 1, ] <- rnorm(n*m1)
x1[, 2, ] <- rgamma(n*m1, shape = 1)
x2[, 1, ] <- rnorm(n*m2)
x2[, 2, ] <- rgamma(n*m2, shape = 1)

#es calculation
es(y = y, x = x1, method = "ens", mean = FALSE)
es(y = y, x = x1, method = "ens", mean = TRUE)

es(y = y, x = x2, method = "mc", mean = FALSE)
es(y = y, x = x2, method = "mc", mean = TRUE)

}
\references{
Gneiting, T., Stanberry, L., Grimit, E., Held, L. and Johnson, N. (2008). Assessing probabilistic forecasts of multivariate quantities, with an application to ensemble predictions of surface winds. Test, 17, 211-264.
}
\author{
David Jobst
}
