% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{logs}
\alias{logs}
\title{Logarithmic Score}
\usage{
logs(y, mean = FALSE)
}
\arguments{
\item{y}{vector of density values}

\item{mean}{logical; if \code{TRUE} the mean of the LogS values is calculated for output; if \code{FALSE} the single LogS values are used as output; default: \code{FALSE}}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Logarithmic Score (LogS) given density values of observations of a uni- or multivariate-dimensional predictive distribution.
}
\details{
For a predictive density function f the density values \code{y} of observations \code{z} are
obtained by \code{y} = f(\code{z}). For a "large" number of ensemble forecasts, the density function(s)
of the ensemble forecasts may be approximated by using, e.g. kernel density
estimation or by fitting a parametric distribution.
Only finite values of \code{y} are used.

A lower LogS indicates a better forecast.
}
\examples{
#simulated data
n <- 30
x <- sample(-10:35, size = n, replace = TRUE)
y <- dnorm(x = x, mean = 0, sd = 1)

#logs calculation
logs(y = y)
logs(y = y, mean = TRUE)

}
\references{
Good, I. (1952). Rational decisions. Journal of the Royal Statistical Society Ser. B, 14, 107-114.
}
\author{
David Jobst
}
