% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{qs}
\alias{qs}
\title{Quantile Score}
\usage{
qs(y, q, p, mean = FALSE, na.rm = FALSE)
}
\arguments{
\item{y}{vector of observations}

\item{q}{vector of quantile values (depending on \code{p}; see details)}

\item{p}{vector of probabilities in (0,1)}

\item{mean}{logical; if \code{TRUE} the mean of the QS values is calculated for output; if \code{FALSE} the single QS values are used as output; default: \code{FALSE}}

\item{na.rm}{logical; if \code{TRUE} NA are removed after the computation; if \code{FALSE} NA are used in the computation; default: \code{FALSE}}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Quantile Score (QS) given observations of a one-dimensional variable, probabilities and corresponding quantiles of a predictive distribution.
}
\details{
For a vector \code{y} of length n, the i-th entry of \code{y} belongs to the i-th entry
of \code{q} and \code{p}. For given probabilities \code{p} the quantile values are obtained by
\code{q}=F^{-1}(\code{p}) for a predictive distribution F.

A lower QS indicates a better forecast.
}
\examples{
#simulated data
n <- 30
y <- rnorm(n)
p <- runif(n)
q <- qnorm(p)

#bs calculation
qs(y = y, q = q, p = p, mean = FALSE)
qs(y = y, q = q, p = p, mean = TRUE)

}
\references{
Friedrich, P. and Hense, A. (2007). Statistical downscaling of extreme precipitation events using censored quantile regression. Monthly Weather Review, 135, 2365-2378.

Gneiting, T. and Raftery, A. (2007). Strictly proper scoring rules, prediction, and estimation. Journal of the American Statistical Association, 102, 359-378.
}
\author{
David Jobst
}
