% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmv.R
\name{rmv}
\alias{rmv}
\title{Root Mean Variance}
\usage{
rmv(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of variances or matrix of ensemble forecasts/samples of a predictive distribution (see details)}

\item{na.rm}{logical; if \code{TRUE} NA are removed after the computation; if \code{FALSE} NA are used in the computation; default: \code{FALSE}}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Root Mean Variance (RMV) given ensemble forecasts/samples of a predictive distribution.
}
\details{
If \code{x} is provided as matrix, it should contain the
ensemble forecasts or the samples of a predictive distribution in each row. Consequently,
the variances are calculated row-wise by its sample version.
If \code{x} is provided as vector, it should contain the variances of the
ensemble forecasts or of a predictive distribution.

A lower RMV indicates a sharper forecast. The optimal value of the RMV is 0.
}
\examples{
#simulated data
n <- 30
m <- 50
x1 <- matrix(rnorm(n*m), ncol = m)
x2 <- apply(x1, 1, var)

#rmv calculation
rmv(x = x1)
rmv(x = x2)

}
\references{
Gneiting, T. and Ranjan, R. (2013). Combining predictive distributions. Electronic Journal of Statistics, 7, 1747-1782.
}
\author{
David Jobst
}
