% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Squared Error}
\usage{
se(y, x, mean = FALSE)
}
\arguments{
\item{y}{vector of observations}

\item{x}{matrix of ensemble forecasts/samples of a predictive distribution or vector of means of ensemble forecasts/a predictive distribution (depending on \code{y}; see details)}

\item{mean}{logical; if \code{TRUE} the mean of the SE values is calculated for output; if \code{FALSE} the single SE values are used as output; default: \code{FALSE}}
}
\value{
Vector of score value(s).
}
\description{
This function calculates the Squared Error (SE) given observations of a one-dimensional variable and ensemble forecasts/samples of a predictive distribution.
}
\details{
For a vector \code{y} of length n, \code{x} can be given as matrix of ensemble forecasts/samples
with n rows, where the i-th entry of \code{y} belongs to the i-th row
of \code{x}. The columns of \code{x} represent the samples of a predictive distribution
or ensemble forecasts. The row-wise means are determined by its sample version.

If the means of ensemble forecasts or of a predictive distribution are directly
available, \code{x} can be given as vector of means, where
the i-th entry of \code{y} belongs to the i-th entry of \code{x}.
Only finite values of \code{y} and \code{x} are used.

In addition, with this function, the Mean Squared Error (MSE) and Root Mean Squares Error (RMSE) can be calculated (see examples).

A lower SE indicates a better forecast.
}
\examples{
#simulated data
n <- 30
m <- 50
y <- rnorm(n)
x1 <- matrix(rnorm(n*m), ncol = m)
x2 <- apply(x1, 1, mean)

#ae calculation
se(y = y, x = x1, mean = FALSE)
se(y = y, x = x1, mean = TRUE)

se(y = y, x = x2, mean = FALSE)
(mse <- se(y = y, x = x2, mean = TRUE))
(rmse <- sqrt(mse))

}
\references{
Vannitsem, S., Wilks, D. and Messner, J. (2018). Statistical Postprocessing of Ensemble Forecasts. Elsevier. 164.
}
\author{
David Jobst
}
