% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.R
\name{vs}
\alias{vs}
\title{Variogram Score}
\usage{
vs(y, x, p = 0.5, w = NULL, na.action = na.omit, aggregate = FALSE, ...)
}
\arguments{
\item{y}{matrix of observations (see details)}

\item{x}{3-dimensional array of samples of a predictive distribution (depending on \code{y}; see details)}

\item{p}{positive value; default: \code{p = 0.5} (see details)}

\item{w}{quadratic matrix with non-negative weights; default: \code{NULL}, i.e. matrix with constant weights 1 (see details)}

\item{na.action}{function to handle the NA's. Default: \code{na.omit}.}

\item{aggregate}{logical or function for aggregating the single scores, e.g. \code{sum}, \code{mean}, \code{weighted.mean}, ....
Default: \code{FALSE}, i.e. no aggregation function.}

\item{...}{further arguments passed to the \code{aggregate} function.}
}
\value{
Vector of the score value(s).
}
\description{
This function calculates the Variogram Score (VS) given observations of a multivariate variable and samples of a predictive distribution.
}
\details{
The observations are given in the matrix \code{y} with n rows and d columns, where each column belongs to an univariate observation variable.
The i-th row of matrix \code{y} belongs to the i-th third dimension entry of the array \code{x}. The i-th third dimension
entry of \code{x} must be a matrix with n rows, having the same structure as \code{y}, filled with samples of a multivariate predictive distribution.
The quadratic weight matrix \code{w} must have d columns and rows.

A lower VS indicates a better forecast.
}
\examples{
# simulated data
n <- 30
m <- 50
y <- cbind(rnorm(n), rgamma(n, shape = 1))
x <- array(NA, dim = c(m, 2, n))
x[, 1, ] <- rnorm(n*m)
x[, 2, ] <- rgamma(n*m, shape = 1)
w <- matrix(c(0.2, 0.3, 0.3, 0.7), ncol = 2, byrow = TRUE)
p <- 2

# vs calculation
vs(y = y, x = x)
vs(y = y, x = x, aggregate = mean)

vs(y = y, x = x, w = w, p = p)
vs(y = y, x = x, w = w, p = p, aggregate = mean)

}
\references{
Scheurer, M. and Hamill, T. (2015). Variogram-based proper scoring rules for probabilistic forecasts of multivariate quantities. Monthly Weather Review, 143, 1321-1334.
}
\author{
David Jobst
}
