% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.R
\name{vs}
\alias{vs}
\title{Variogram Score}
\usage{
vs(y, x, w = NULL, p = NULL, mean = FALSE)
}
\arguments{
\item{y}{matrix of observations (see details)}

\item{x}{3-dimensional array of ensemble forecasts/samples of a predictive distribution (depending on \code{y}; see details)}

\item{w}{3-dimensional array of quadratic matrices with nonnegative weights; default: matrices with constant weight 1 (see details)}

\item{p}{vector with positive values; default: vector with constant value 0.5 (see details)}

\item{mean}{logical; if \code{TRUE} the mean of the VS values is calculated for output; if \code{FALSE} the single VS values are used as output; default: \code{FALSE}}
}
\value{
Vector of the score value(s).
}
\description{
This function calculates the Variogram Score (VS) given observations of a multivariate variable and ensemble forecasts/samples of a predictive distribution.
}
\details{
The observations are given in the matrix \code{y} with n rows and d columns, where each column belongs to an univariate observation variable.
The i-th row of matrix \code{y} belongs to the i-th third dimension entry of the array \code{x} and \code{w} and to the i-th entry of \code{p}. The i-th third dimension
entry of \code{x} must be a matrix with n rows, having the same structure as \code{y}, filled with the ensemble forecasts or samples of a predictive distribution.
Only finite values of \code{y}, \code{x}, \code{w} and \code{p} are used.

The 3-dimensional array \code{w} must contain quadratic matrices of dimension d with nonnegative weights.

The vector \code{p} of length n must contain positive values

A lower VS indicates a better forecast.
}
\examples{
#simulated data
n <- 30
m <- 50
y <- cbind(rnorm(n), rgamma(n, shape = 1))
x <- array(NA, dim = c(m, 2, n))
x[, 1, ] <- rnorm(n*m)
x[, 2, ] <- rgamma(n*m, shape = 1)
w <- array(NA, dim = c(2, 2, n))
for (i in 1:n) {
w[, , i] <- matrix(runif(4), ncol = 2)
}
p <- runif(n)

#vs calculation
vs(y = y, x = x, mean = FALSE)
vs(y = y, x = x, mean = TRUE)

vs(y = y, x = x, w = w, p = p, mean = FALSE)
vs(y = y, x = x, w = w, p = p, mean = TRUE)

}
\references{
Scheurer, M. and Hamill, T. (2015). Variogram-based proper scoring rules for probabilistic forecasts of multivariate quantities. Monthly Weather Review, 143, 1321-1334.
}
\author{
David Jobst
}
